Embedd Reader in Writer and fixes
