+++ chainio/eth/reader.go
@@ -544,69 +544,3 @@ func (t *Reader) updateContractBindings(blsOperatorStateRetrieverAddr, eigenDASe
 	}
 	return nil
 }
-
-func signatureToBN254G1Point(s *bn254.Signature) eigendasrvmg.BN254G1Point {
-	return eigendasrvmg.BN254G1Point{
-		X: s.X.BigInt(new(big.Int)),
-		Y: s.Y.BigInt(new(big.Int)),
-	}
-}
-
-func pubKeyG1ToBN254G1Point(p *bn254.G1Point) eigendasrvmg.BN254G1Point {
-	return eigendasrvmg.BN254G1Point{
-		X: p.X.BigInt(new(big.Int)),
-		Y: p.Y.BigInt(new(big.Int)),
-	}
-}
-
-func pubKeyG2ToBN254G2Point(p *bn254.G2Point) eigendasrvmg.BN254G2Point {
-	return eigendasrvmg.BN254G2Point{
-		X: [2]*big.Int{p.X.A1.BigInt(new(big.Int)), p.X.A0.BigInt(new(big.Int))},
-		Y: [2]*big.Int{p.Y.A1.BigInt(new(big.Int)), p.Y.A0.BigInt(new(big.Int))},
-	}
-}
-
-func quorumIDsToQuorumNumbers(quorumIds []chainio.QuorumID) []byte {
-	quorumNumbers := make([]byte, len(quorumIds))
-	for i, quorumId := range quorumIds {
-		quorumNumbers[i] = byte(quorumId)
-	}
-	return quorumNumbers
-}
-
-func HashPubKeyG1(pk *bn254.G1Point) [32]byte {
-	gp := pubKeyG1ToBN254G1Point(pk)
-	xBytes := make([]byte, 32)
-	yBytes := make([]byte, 32)
-	gp.X.FillBytes(xBytes)
-	gp.Y.FillBytes(yBytes)
-	return crypto.Keccak256Hash(append(xBytes, yBytes...))
-}
-
-func BitmapToQuorumIds(bitmap *big.Int) []chainio.QuorumID {
-	// loop through each index in the bitmap to construct the array
-
-	quorumIds := make([]chainio.QuorumID, 0, maxNumberOfQuorums)
-	for i := 0; i < maxNumberOfQuorums; i++ {
-		if bitmap.Bit(i) == 1 {
-			quorumIds = append(quorumIds, chainio.QuorumID(i))
-		}
-	}
-	return quorumIds
-}
-
-func bitmapToBytesArray(bitmap *big.Int) []byte {
-	// initialize an empty uint64 to be used as a bitmask inside the loop
-	var (
-		bytesArray []byte
-	)
-	// loop through each index in the bitmap to construct the array
-	for i := 0; i < maxNumberOfQuorums; i++ {
-		// check if the i-th bit is flipped in the bitmap
-		if bitmap.Bit(i) == 1 {
-			// if the i-th bit is flipped, then add a byte encoding the value 'i' to the `bytesArray`
-			bytesArray = append(bytesArray, byte(uint8(i)))
-		}
-	}
-	return bytesArray
-}
