+++ chainio/mock/writer.go
@@ -13,6 +13,7 @@ import (
 )
 
 type MockWriter struct {
+	MockReader
 	mock.Mock
 }
